Backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
